+++ src/Tools/PDF/FooterTemplate.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LBF\Tools\PDF;
+
+interface FooterTemplate {
+
+    public function set_footer(): void;
+}