+++ src/Tools/PDF/HeaderTemplate.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LBF\Tools\PDF;
+
+interface HeaderTemplate {
+
+    public function set_header(): void;
+}